Update private.py
Hello. I'm using your code to access to Crypto Exchange. While testing my code i realize that when i was trying to buy CRO with BTC, there was en error.
In function create_order, you were modifying 'quantity' always with the quantity precision, so when I wanted to sell BTC to buy CRO, the quantity passed was 0.00 and there was an exception.
In fact, when we sell BTC to buy CRO, we have to use 'notional', so i changed the way you modify the precision with price precision. Now it works.